+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@
 ### 🗑 Deprecations
 
 * `ImageRenderer` was not available to 5.7 compilers and has been removed altogether.
+* `ImageRenderer` can still be found and copied from any earlier versions of SwiftUIKit.
 * `SystemColor` is deprecated.