+++ RELEASE_NOTES.md
@@ -18,6 +18,28 @@
 
 ### 🗑 Deprecations
 
+* `AlertProvider` has been deprecated.
+* `FullScreenCoverProvider` has been deprecated.
+* `SheetProvider` has been deprecated.
+
+
+## 3.4
+
+### ✨ New features
+
+* `Color+Hex` has new static functions.
+* `ColorPickerBar` is a new color picker component.
+* `EditableView` is a new view protocol.
+* `ListBadgeIcon` now supports custom icon colors and icon gradients.
+* `ListHeaderIcon` is a new view.
+* `ListSectionTitle` has a new `withInsets()` init parameter. 
+* `ListShelfSection` has a new style.
+* `OptionalBinding` is a new func that makes it possible to pass in optional values as a non-optional binding.
+* `View` has new `prefers` extensions.
+* `ViewShadowStyle` has a new `.sticker` style.
+
+### 🗑 Deprecations
+
 * `ImageRenderer` was not available to 5.7 compilers and has been removed altogether.
 * `ImageRenderer` can still be found and copied from any earlier versions of SwiftUIKit.
 * `MultilineTextField` has been deprecated.